requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
Flask
PyYAML
pydantic
nltk
schedule
python-dotenv